#209 Refactor TableDefProcessor — extract cohesive collaborator classes
Pure refactor with no behavioural change. Splits the 2120-LOC god class into
five focused collaborators in the same package:

- TableDefConstants.kt — shared FQN constants and toSnakeCase() extension
- TableDefMetaTypes.kt — all data classes (ColumnMeta, AggregatePropertyMeta,
  JunctionRefInfo, ForeignKeyMeta, PersistencePropertyHints, ConverterInfo)
- TableDefSourceEmitter.kt — all StringBuilder.append* and build* helper
  functions; pure string-building with no KSP model dependency
- ColumnMetaBuilder.kt — per-property ColumnMeta resolution, converter
  validation, type-expression derivation, and property exclusion rules
- EmbeddableAnalyzer.kt — @Embedded/@Embeddable recursive descent,
  D-06/D-07 validation, collectColumnsAndSlots, detectColumnCollisions
- ForeignKeyAnalyzer.kt — @Aggregate FK and junction-table analysis,
  FK-04/FK-05 validation, generateJunctionTableDef

TableDefProcessor retains only the KSP SymbolProcessor orchestration:
process(), aggregate scanning, class collection, version validation, and
generateTableDef(); all other logic delegates to the new collaborators.
