+++ lirp-ksp/src/main/kotlin/net/transgressoft/lirp/ksp/TableDefProcessor.kt
@@ -17,95 +17,22 @@
 
 package net.transgressoft.lirp.ksp
 
-import com.google.devtools.ksp.getDeclaredProperties
-import com.google.devtools.ksp.isAbstract
+import net.transgressoft.lirp.ksp.TableDefSourceEmitter.appendObjectBody
+import net.transgressoft.lirp.ksp.TableDefSourceEmitter.appendPackageAndImports
 import com.google.devtools.ksp.processing.CodeGenerator
 import com.google.devtools.ksp.processing.Dependencies
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.processing.SymbolProcessor
-import com.google.devtools.ksp.symbol.ClassKind
 import com.google.devtools.ksp.symbol.FileLocation
 import com.google.devtools.ksp.symbol.KSAnnotated
-import com.google.devtools.ksp.symbol.KSAnnotation
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSPropertyDeclaration
 import com.google.devtools.ksp.symbol.KSType
-import com.google.devtools.ksp.symbol.KSValueParameter
-import com.google.devtools.ksp.symbol.Modifier
-import com.google.devtools.ksp.symbol.Origin
 import com.google.devtools.ksp.validate
 import java.io.File
 
-private const val PERSISTENCE_MAPPING_FQN = "net.transgressoft.lirp.persistence.PersistenceMapping"
-private const val PERSISTENCE_PROPERTY_FQN = "net.transgressoft.lirp.persistence.PersistenceProperty"
-private const val PERSISTENCE_IGNORE_FQN = "net.transgressoft.lirp.persistence.PersistenceIgnore"
 private const val AGGREGATE_ANNOTATION_FQN = "net.transgressoft.lirp.persistence.Aggregate"
-private const val VERSION_FQN = "net.transgressoft.lirp.persistence.Version"
-private const val TRANSIENT_FQN = "kotlin.jvm.Transient"
-private const val SQL_TABLE_DEF_FQN = "net.transgressoft.lirp.persistence.sql.SqlTableDef"
-private const val UUID_FQN = "java.util.UUID"
-private const val LOCAL_DATE_FQN = "java.time.LocalDate"
-private const val LOCAL_DATE_TIME_FQN = "java.time.LocalDateTime"
-private const val KOTLIN_STRING_FQN = "kotlin.String"
-private const val KOTLIN_INT_FQN = "kotlin.Int"
-private const val KOTLIN_LONG_FQN = "kotlin.Long"
-private const val KOTLIN_SHORT_FQN = "kotlin.Short"
-private const val KOTLIN_BYTE_FQN = "kotlin.Byte"
-private const val KOTLIN_BOOLEAN_FQN = "kotlin.Boolean"
-private const val KOTLIN_DOUBLE_FQN = "kotlin.Double"
-private const val KOTLIN_FLOAT_FQN = "kotlin.Float"
-private const val KOTLIN_UUID_FQN = "kotlin.UUID"
-private const val BIG_DECIMAL_FQN = "java.math.BigDecimal"
-private const val COLUMN_TYPE_INT_EXPR = "ColumnType.IntType"
-private const val COLUMN_TYPE_TEXT_EXPR = "ColumnType.TextType"
-private const val COLUMN_TYPE_LONG_EXPR = "ColumnType.LongType"
-private const val COLUMN_TYPE_BOOLEAN_EXPR = "ColumnType.BooleanType"
-private const val COLUMN_TYPE_DOUBLE_EXPR = "ColumnType.DoubleType"
-private const val COLUMN_TYPE_FLOAT_EXPR = "ColumnType.FloatType"
-private const val COLUMN_TYPE_UUID_EXPR = "ColumnType.UuidType"
-private const val COLUMN_TYPE_DATE_EXPR = "ColumnType.DateType"
-private const val COLUMN_TYPE_DATETIME_EXPR = "ColumnType.DateTimeType"
-private const val LIST_ITEM_SEPARATOR = ",\n        "
-private const val COLUMN_CONVERTER_FQN = "net.transgressoft.lirp.persistence.ColumnConverter"
-private const val EMBEDDABLE_FQN = "net.transgressoft.lirp.persistence.Embeddable"
-private const val EMBEDDED_FQN = "net.transgressoft.lirp.persistence.Embedded"
-
-// Allow-list of supported S type FQNs for ColumnConverter<D, S>, mapped to the canonical
-// ColumnType expression. Keys gate D-08 validation; values seed the codegen path that
-// emits the column type from the converter's base scalar.
-private val SUPPORTED_CONVERTER_S_TYPES: Map<String, String> =
-    mapOf(
-        KOTLIN_STRING_FQN to COLUMN_TYPE_TEXT_EXPR,
-        KOTLIN_INT_FQN to COLUMN_TYPE_INT_EXPR,
-        KOTLIN_LONG_FQN to COLUMN_TYPE_LONG_EXPR,
-        KOTLIN_SHORT_FQN to COLUMN_TYPE_INT_EXPR,
-        KOTLIN_BYTE_FQN to COLUMN_TYPE_INT_EXPR,
-        KOTLIN_BOOLEAN_FQN to COLUMN_TYPE_BOOLEAN_EXPR,
-        KOTLIN_DOUBLE_FQN to COLUMN_TYPE_DOUBLE_EXPR,
-        KOTLIN_FLOAT_FQN to COLUMN_TYPE_FLOAT_EXPR,
-        BIG_DECIMAL_FQN to "ColumnType.DecimalType",
-        UUID_FQN to COLUMN_TYPE_UUID_EXPR,
-        LOCAL_DATE_FQN to COLUMN_TYPE_DATE_EXPR,
-        LOCAL_DATE_TIME_FQN to COLUMN_TYPE_DATETIME_EXPR
-    )
-
-internal data class ConverterInfo(
-    val converterFqn: String,
-    val sqlTypeFqn: String
-)
-
-/**
- * Bundle of `@PersistenceProperty` hint values consumed by the codegen path that derives
- * a column's `typeExpression`. Carrying them as a single value keeps refinement-helper
- * signatures focused — see [TableDefProcessor.refineConverterSqlType].
- */
-internal data class PersistencePropertyHints(
-    val length: Int,
-    val precision: Int,
-    val scale: Int,
-    val typeHint: String
-)
 
 /**
  * KSP processor that generates `_LirpTableDef` descriptor objects for entity classes annotated with
@@ -130,11 +57,19 @@ internal data class PersistencePropertyHints(
  *
  * Both annotation entry points are supported: a class-level `@PersistenceMapping` and a property-level
  * `@PersistenceProperty` on a class without `@PersistenceMapping` both trigger generation.
+ *
+ * Actual code generation is delegated to [TableDefSourceEmitter] (string building), [EmbeddableAnalyzer]
+ * (`@Embedded` / `@Embeddable` recursive descent), [ForeignKeyAnalyzer] (FK constraints and junction
+ * tables), and [ColumnMetaBuilder] (per-property column resolution).
  */
 class TableDefProcessor(
     private val codeGenerator: CodeGenerator,
     private val logger: KSPLogger
 ) : SymbolProcessor {
+
+    private val columnMetaBuilder = ColumnMetaBuilder(logger)
+    private val embeddableAnalyzer = EmbeddableAnalyzer(logger, columnMetaBuilder)
+    private val foreignKeyAnalyzer = ForeignKeyAnalyzer(logger, codeGenerator, columnMetaBuilder)
 
     override fun process(resolver: Resolver): List<KSAnnotated> {
         val unableToProcess = mutableListOf<KSAnnotated>()
@@ -153,9 +88,9 @@ class TableDefProcessor(
 
         for (classDecl in classes) {
             val aggregates = aggregatesByClass[classDecl].orEmpty()
-            val foreignKeys = sqlTableDefAvailable.let { collectForeignKeys(classDecl, aggregates) }
+            val foreignKeys = sqlTableDefAvailable.let { foreignKeyAnalyzer.collectForeignKeys(classDecl, aggregates) }
             val junctionRefs =
-                if (sqlTableDefAvailable) collectJunctionRefs(classDecl, aggregates) else emptyList()
+                if (sqlTableDefAvailable) foreignKeyAnalyzer.collectJunctionRefs(classDecl, aggregates) else emptyList()
             // Backing collection fields for @Aggregate collection refs are never SQL columns —
             // they are the in-memory mirror of junction-table rows. Exclude them from column
             // collection so the column-type mapper doesn't emit "Unsupported column type
@@ -179,7 +114,7 @@ class TableDefProcessor(
 
             if (sqlTableDefAvailable) {
                 for (collectionAgg in aggregates.filter { it.isCollection }) {
-                    generateJunctionTableDef(classDecl, collectionAgg)
+                    foreignKeyAnalyzer.generateJunctionTableDef(classDecl, collectionAgg)
                 }
             }
         }
@@ -219,7 +154,7 @@ class TableDefProcessor(
             prop.annotations.firstOrNull {
                 it.annotationType.resolve().declaration.qualifiedName?.asString() == AGGREGATE_ANNOTATION_FQN
             } ?: return null
-        val onDeleteName = extractCascadeActionName(annotation.arguments.firstOrNull { it.name?.asString() == "onDelete" }?.value)
+        val onDeleteName = foreignKeyAnalyzer.extractCascadeActionName(annotation.arguments.firstOrNull { it.name?.asString() == "onDelete" }?.value)
 
         val sourceText = readSourceLines(prop, linesAfter = 2) ?: ""
         val isCollection = containsAggregateFactoryCall(sourceText)
@@ -255,7 +190,8 @@ class TableDefProcessor(
      * `aggregateList(initialTrackIds)` / `aggregateSet(initialTagIds)` (and their `mutable*`
      * variants). Returns `null` when the argument is not a bare identifier (e.g. `emptyList()`,
      * `setOf(1, 2)`), which signals that the entity has no writable backing field for the
-     * collection — the validation in [collectJunctionRefs] surfaces a clear error in that case.
+     * collection — the validation in [ForeignKeyAnalyzer.collectJunctionRefs] surfaces a clear
+     * error in that case.
      */
     private fun extractAggregateCollectionArgument(text: String): String? {
         // Match `aggregateList<...>(IDENT)` / `aggregateSet<...>(IDENT)` and their mutable
@@ -474,7 +410,7 @@ class TableDefProcessor(
 
         val tableName = resolveTableName(classDecl, className)
         val collected =
-            collectColumnsAndSlots(classDecl, versionedProperty, excludedBackingFields, aggregateBackingScalarNames)
+            embeddableAnalyzer.collectColumnsAndSlots(classDecl, versionedProperty, excludedBackingFields, aggregateBackingScalarNames)
         val columns = collected.columns
         val ctorSlots = collected.ctorSlots
 
@@ -482,7 +418,7 @@ class TableDefProcessor(
         // flattened column list, after all recursive @Embedded descents. Detection at this level
         // (rather than per-@Embeddable) catches grandchild collisions that an intermediate prefix
         // might otherwise mask. Codegen is suppressed for the entity when collisions are reported.
-        if (detectColumnCollisions(classDecl, columns)) {
+        if (embeddableAnalyzer.detectColumnCollisions(classDecl, columns)) {
             logger.info("Skipping _LirpTableDef generation for $className due to column-name collisions")
             return
         }
@@ -538,401 +474,6 @@ class TableDefProcessor(
         logger.info("Generated $packageName.$tableDefName for $className (sqlTableDef=$canGenerateSqlMapping)")
     }
 
-    private fun StringBuilder.appendPackageAndImports(
-        packageName: String,
-        canGenerateSqlMapping: Boolean,
-        columns: List<ColumnMeta>,
-        emitsForeignKeys: Boolean = false,
-        emitsJunctions: Boolean = false
-    ) {
-        if (packageName.isNotEmpty()) {
-            appendLine("package $packageName")
-            appendLine()
-        }
-        appendLine("import net.transgressoft.lirp.persistence.ColumnDef")
-        appendLine("import net.transgressoft.lirp.persistence.ColumnType")
-        if (canGenerateSqlMapping) {
-            appendLine("import net.transgressoft.lirp.persistence.LirpTableDef")
-            appendLine("import net.transgressoft.lirp.persistence.sql.SqlTableDef")
-            appendLine("import org.jetbrains.exposed.v1.core.Column")
-            appendLine("import org.jetbrains.exposed.v1.core.ResultRow")
-            appendLine("import org.jetbrains.exposed.v1.core.Table")
-            if (emitsForeignKeys) {
-                appendLine("import net.transgressoft.lirp.entity.CascadeAction")
-                appendLine("import net.transgressoft.lirp.persistence.sql.ForeignKeyDef")
-            }
-            if (emitsJunctions) {
-                appendLine("import net.transgressoft.lirp.persistence.sql.JunctionAccessor")
-                appendLine("import net.transgressoft.lirp.persistence.sql.JunctionTableDef")
-            }
-            appendConditionalTypeImports(columns)
-        } else {
-            appendLine("import net.transgressoft.lirp.persistence.LirpTableDef")
-        }
-        appendLine()
-    }
-
-    private fun StringBuilder.appendConditionalTypeImports(columns: List<ColumnMeta>) {
-        if (columns.any { it.typeFqn == UUID_FQN }) {
-            appendLine("import kotlin.uuid.ExperimentalUuidApi")
-            appendLine("import kotlin.uuid.toJavaUuid")
-            appendLine("import kotlin.uuid.toKotlinUuid")
-        }
-        if (columns.any { it.typeFqn == LOCAL_DATE_FQN }) {
-            appendLine("import kotlinx.datetime.toJavaLocalDate")
-            appendLine("import kotlinx.datetime.toKotlinLocalDate")
-        }
-        if (columns.any { it.typeFqn == LOCAL_DATE_TIME_FQN }) {
-            appendLine("import kotlinx.datetime.toJavaLocalDateTime")
-            appendLine("import kotlinx.datetime.toKotlinLocalDateTime")
-        }
-        if (columns.any { it.typeExpression.startsWith("ColumnType.DecimalType") }) {
-            appendLine("import java.math.BigDecimal")
-        }
-        columns.filter { it.isEnum }.map { it.typeFqn }.distinct().forEach { fqn ->
-            appendLine("import $fqn")
-        }
-    }
-
-    /** Aggregates the descriptor-shape inputs the table-def emitter consumes, keeping the
-     *  emitter API below Sonar's 7-parameter ceiling without inlining the structure at every
-     *  call site. */
-    private data class ObjectBodyParams(
-        val tableName: String,
-        val canGenerateSqlMapping: Boolean,
-        val columns: List<ColumnMeta>,
-        val constructorParamNames: List<String> = emptyList(),
-        val ctorSlots: List<CtorSlot> = emptyList(),
-        val foreignKeys: List<ForeignKeyMeta> = emptyList(),
-        val junctionRefs: List<JunctionRefInfo> = emptyList()
-    )
-
-    private fun StringBuilder.appendObjectBody(
-        tableDefName: String,
-        className: String,
-        params: ObjectBodyParams
-    ) {
-        val tableName = params.tableName
-        val canGenerateSqlMapping = params.canGenerateSqlMapping
-        val columns = params.columns
-        val constructorParamNames = params.constructorParamNames
-        val foreignKeys = params.foreignKeys
-        val junctionRefs = params.junctionRefs
-        appendLine("/** KSP-generated table descriptor for [$className]. */")
-        if (canGenerateSqlMapping && columns.any { it.typeFqn == UUID_FQN }) {
-            appendLine("@OptIn(ExperimentalUuidApi::class)")
-        }
-        val superType = if (canGenerateSqlMapping) "SqlTableDef<$className>" else "LirpTableDef<$className>"
-        appendLine("public object $tableDefName : $superType {")
-        appendLine("    override val tableName: String = \"$tableName\"")
-        appendLine("    override val columns: List<ColumnDef> = listOf(")
-        if (columns.isNotEmpty()) {
-            val columnsCode =
-                columns.joinToString(LIST_ITEM_SEPARATOR) { col ->
-                    "ColumnDef(name = \"${col.name}\", type = ${col.typeExpression}, " +
-                        "nullable = ${col.nullable}, primaryKey = ${col.isPrimaryKey}, isVersion = ${col.isVersion})"
-                }
-            appendLine("        $columnsCode")
-        }
-        appendLine("    )")
-        if (canGenerateSqlMapping) {
-            appendLine()
-            appendFromRow(className, columns, constructorParamNames, params.ctorSlots)
-            appendLine()
-            appendToParams(className, columns)
-            appendLine()
-            appendApplyRow(className, columns)
-            appendLine()
-            appendApplyScalarRow(className, columns)
-            appendBumpVersion(className, columns)
-            appendForeignKeys(foreignKeys)
-            appendJunctionOverrides(className, junctionRefs)
-        }
-        appendLine("}")
-    }
-
-    private fun StringBuilder.appendJunctionOverrides(
-        className: String,
-        junctionRefs: List<JunctionRefInfo>
-    ) {
-        if (junctionRefs.isEmpty()) return
-        appendLine()
-        appendLine("    override val junctionTableDefs: List<JunctionTableDef> = listOf(")
-        appendLine("        ${junctionRefs.joinToString(LIST_ITEM_SEPARATOR) { it.junctionObjectName }}")
-        appendLine("    )")
-        appendLine()
-        appendLine("    override val junctionAccessors: List<JunctionAccessor<$className>> = listOf(")
-        junctionRefs.forEachIndexed { idx, ref ->
-            val trailingComma = if (idx == junctionRefs.lastIndex) "" else ","
-            appendLine("        object : JunctionAccessor<$className> {")
-            appendLine("            override val descriptor: JunctionTableDef = ${ref.junctionObjectName}")
-            appendLine("            override fun idsOf(entity: $className): Collection<Any> = entity.${ref.backingFieldName}")
-            appendLine("        }$trailingComma")
-        }
-        appendLine("    )")
-        appendLine()
-        appendLine("    override fun applyJunctionRows(")
-        appendLine("        entity: $className,")
-        appendLine("        descriptor: JunctionTableDef,")
-        appendLine("        ids: List<Any>,")
-        appendLine("    ) {")
-        appendLine("        entity.withEventsDisabled {")
-        appendLine("            when (descriptor) {")
-        for (ref in junctionRefs) {
-            val rhs =
-                if (ref.isOrdered) {
-                    "ids.filterIsInstance<${ref.itemKeyTypeSimpleName}>()"
-                } else {
-                    "ids.filterIsInstance<${ref.itemKeyTypeSimpleName}>().toSet()"
-                }
-            appendLine("                ${ref.junctionObjectName} ->")
-            appendLine("                    entity.${ref.backingFieldName} = $rhs")
-        }
-        appendLine("            }")
-        appendLine("        }")
-        appendLine("    }")
-    }
-
-    private fun StringBuilder.appendForeignKeys(foreignKeys: List<ForeignKeyMeta>) {
-        if (foreignKeys.isEmpty()) return
-        appendLine()
-        appendLine("    override fun foreignKeys(): List<ForeignKeyDef> = listOf(")
-        val entries =
-            foreignKeys.joinToString(LIST_ITEM_SEPARATOR) { fk ->
-                "ForeignKeyDef(columnName = \"${fk.columnName}\", " +
-                    "referencedTable = \"${fk.referencedTable}\", " +
-                    "referencedColumn = \"${fk.referencedColumn}\", " +
-                    "onDelete = CascadeAction.${fk.onDelete})"
-            }
-        appendLine("        $entries")
-        appendLine("    )")
-    }
-
-    private fun StringBuilder.appendFromRow(
-        className: String,
-        columns: List<ColumnMeta>,
-        constructorParamNames: List<String>,
-        ctorSlots: List<CtorSlot> = emptyList()
-    ) {
-        val columnsByName = columns.associateBy { it.propertyName }
-        // Preserve constructor parameter declaration order for correct positional arguments
-        val orderedCtorCols = constructorParamNames.mapNotNull { columnsByName[it] }
-        val ctorParamNameSet = constructorParamNames.toSet()
-        // Setter cols exclude both ctor params and embedded-derived columns (the latter share their
-        // top-level entity ctor-param name and are reconstructed inside the ctor invocation).
-        val setterCols = columns.filter { it.propertyName !in ctorParamNameSet && !it.isInsideEmbedded }
-
-        appendLine("    override fun fromRow(row: ResultRow, table: Table): $className {")
-        // When CtorSlot information is available (entities with @Embedded ctor params), use the
-        // structured tree to emit nested constructor expressions; otherwise fall back to the flat
-        // column-by-name lookup for the common no-embedded case.
-        val ctorArgs =
-            if (ctorSlots.isNotEmpty()) {
-                ctorSlots.joinToString(", ") { buildCtorArgExpression(it) }
-            } else {
-                orderedCtorCols.joinToString(", ") { buildRowAccess(it) }
-            }
-        appendLine("        val entity = $className($ctorArgs)")
-        for (col in setterCols) {
-            val rowAccess = buildRowAccess(col)
-            appendLine("        entity.${col.propertyName} = $rowAccess")
-        }
-        appendLine("        return entity")
-        appendLine("    }")
-    }
-
-    /**
-     * Recursively emits a Kotlin expression that reconstructs the value bound to [slot] from the
-     * current `ResultRow`. Scalar slots delegate to [buildRowAccess]; embedded slots emit a nested
-     * constructor invocation whose arguments are themselves recursive expressions.
-     */
-    private fun buildCtorArgExpression(slot: CtorSlot): String =
-        when (slot) {
-            is ScalarCtorSlot -> buildRowAccess(slot.column)
-            is EmbeddedCtorSlot -> {
-                val inner =
-                    slot.children.joinToString(", ") { child ->
-                        "${child.ctorParamName} = ${buildCtorArgExpression(child)}"
-                    }
-                "${slot.embeddableTypeFqn}($inner)"
-            }
-        }
-
-    private fun buildRowAccess(col: ColumnMeta): String {
-        val rawAccess = "row[table.columns.first { it.name == \"${col.name}\" }]"
-        return buildConverterRowAccess(col, rawAccess)
-            ?: buildNarrowingIntRowAccess(col, rawAccess)
-            ?: buildBuiltInRowAccess(col, rawAccess)
-    }
-
-    // Converter-routed columns short-circuit the FQN-driven cast table: read the raw scalar,
-    // cast to the converter's S type, then route through the consumer's fromSql. Short/Byte
-    // converters need the same INT → narrow-typed conversion that non-converter columns get
-    // (the JDBC layer boxes the column value as Int regardless of declared width), or the cast
-    // would throw ClassCastException at row time.
-    private fun buildConverterRowAccess(col: ColumnMeta, rawAccess: String): String? {
-        if (col.converterFqn == null || col.converterSqlFqn == null) return null
-        val converterInput =
-            when (col.converterSqlFqn) {
-                KOTLIN_SHORT_FQN ->
-                    if (col.nullable) "($rawAccess as? Int)?.toShort()" else "($rawAccess as Int).toShort()"
-                KOTLIN_BYTE_FQN ->
-                    if (col.nullable) "($rawAccess as? Int)?.toByte()" else "($rawAccess as Int).toByte()"
-                else ->
-                    if (col.nullable) "($rawAccess as? ${col.converterSqlFqn})" else "($rawAccess as ${col.converterSqlFqn})"
-            }
-        return if (col.nullable) {
-            "$converterInput?.let { ${col.converterFqn}.fromSql(it) }"
-        } else {
-            "${col.converterFqn}.fromSql($converterInput)"
-        }
-    }
-
-    // Short / Byte are stored as INT; narrow on read via Kotlin's truncating conversion.
-    // A raw `as Short` would fail at runtime because the JDBC value is boxed as Int.
-    private fun buildNarrowingIntRowAccess(col: ColumnMeta, rawAccess: String): String? =
-        when (col.typeFqn) {
-            KOTLIN_SHORT_FQN ->
-                if (col.nullable) "($rawAccess as? Int)?.toShort()" else "($rawAccess as Int).toShort()"
-            KOTLIN_BYTE_FQN ->
-                if (col.nullable) "($rawAccess as? Int)?.toByte()" else "($rawAccess as Int).toByte()"
-            else -> null
-        }
-
-    private fun buildBuiltInRowAccess(col: ColumnMeta, rawAccess: String): String =
-        when {
-            col.typeFqn == UUID_FQN && col.nullable -> "($rawAccess as? kotlin.uuid.Uuid)?.toJavaUuid()"
-            col.typeFqn == UUID_FQN -> "($rawAccess as kotlin.uuid.Uuid).toJavaUuid()"
-            col.typeFqn == LOCAL_DATE_FQN && col.nullable -> "($rawAccess as? kotlinx.datetime.LocalDate)?.toJavaLocalDate()"
-            col.typeFqn == LOCAL_DATE_FQN -> "($rawAccess as kotlinx.datetime.LocalDate).toJavaLocalDate()"
-            col.typeFqn == LOCAL_DATE_TIME_FQN && col.nullable -> "($rawAccess as? kotlinx.datetime.LocalDateTime)?.toJavaLocalDateTime()"
-            col.typeFqn == LOCAL_DATE_TIME_FQN -> "($rawAccess as kotlinx.datetime.LocalDateTime).toJavaLocalDateTime()"
-            col.isEnum && col.nullable -> {
-                val enumSimpleName = col.typeFqn.substringAfterLast(".")
-                "($rawAccess as? String)?.let { enumValueOf<$enumSimpleName>(it) }"
-            }
-            col.isEnum -> {
-                val enumSimpleName = col.typeFqn.substringAfterLast(".")
-                "enumValueOf<$enumSimpleName>($rawAccess as String)"
-            }
-            col.nullable -> "$rawAccess as? ${col.typeFqn.substringAfterLast(".")}"
-            else -> "$rawAccess as ${col.typeFqn.substringAfterLast(".")}"
-        }
-
-    private fun StringBuilder.appendToParams(className: String, columns: List<ColumnMeta>) {
-        appendLine("    override fun toParams(entity: $className, table: Table): Map<Column<*>, Any?> {")
-        appendLine("        val cols = table.columns.associateBy { it.name }")
-        appendLine("        return mapOf(")
-        val paramEntries =
-            columns.joinToString(",\n            ") { col ->
-                val valueAccess = buildEntityAccess(col)
-                "cols[\"${col.name}\"]!! to $valueAccess"
-            }
-        appendLine("            $paramEntries")
-        appendLine("        )")
-        appendLine("    }")
-    }
-
-    private fun buildEntityAccess(col: ColumnMeta): String {
-        // Use embeddedPath so @Embedded-derived columns dereference the nested access path
-        // (e.g. "entity.album.performer.name") rather than the top-level entity ctor-param name.
-        // For non-embedded columns embeddedPath defaults to propertyName so behaviour is unchanged.
-        val prop = "entity.${col.embeddedPath}"
-        // Converter-routed columns short-circuit the FQN-driven write table: route the domain
-        // value through toSql before binding. Short/Byte converters then widen to Int to match
-        // the IntType column the JDBC layer expects (symmetric to buildConverterRowAccess).
-        if (col.converterFqn != null) {
-            val sqlValue =
-                if (col.nullable) "$prop?.let { ${col.converterFqn}.toSql(it) }"
-                else "${col.converterFqn}.toSql($prop)"
-            return when (col.converterSqlFqn) {
-                KOTLIN_SHORT_FQN, KOTLIN_BYTE_FQN ->
-                    if (col.nullable) "$sqlValue?.toInt()" else "$sqlValue.toInt()"
-                else -> sqlValue
-            }
-        }
-        // Short / Byte are stored as INT — widen on write so the bound parameter matches the column.
-        if (col.typeFqn == KOTLIN_SHORT_FQN) {
-            return if (col.nullable) "$prop?.toInt()" else "$prop.toInt()"
-        }
-        if (col.typeFqn == KOTLIN_BYTE_FQN) {
-            return if (col.nullable) "$prop?.toInt()" else "$prop.toInt()"
-        }
-        return when {
-            col.typeFqn == UUID_FQN && col.nullable -> "$prop?.toKotlinUuid()"
-            col.typeFqn == UUID_FQN -> "$prop.toKotlinUuid()"
-            col.typeFqn == LOCAL_DATE_FQN && col.nullable -> "$prop?.toKotlinLocalDate()"
-            col.typeFqn == LOCAL_DATE_FQN -> "$prop.toKotlinLocalDate()"
-            col.typeFqn == LOCAL_DATE_TIME_FQN && col.nullable -> "$prop?.toKotlinLocalDateTime()"
-            col.typeFqn == LOCAL_DATE_TIME_FQN -> "$prop.toKotlinLocalDateTime()"
-            col.isEnum && col.nullable -> "$prop?.name"
-            col.isEnum -> "$prop.name"
-            else -> prop
-        }
-    }
-
-    private fun StringBuilder.appendApplyRow(className: String, columns: List<ColumnMeta>) {
-        // applyRow overwrites the state of an existing entity — skip primary-key columns
-        // (they are immutable post-construction), any non-mutable property, and any column
-        // produced by flattening an @Embedded value object (embeddables are reconstructed
-        // wholesale via the primary constructor in fromRow; the parent entity has no
-        // addressable scalar setter for them).
-        val mutableNonPk = columns.filter { !it.isPrimaryKey && it.isMutable && !it.isInsideEmbedded }
-        appendLine("    override fun applyRow(entity: $className, row: ResultRow, table: Table) {")
-        if (mutableNonPk.isEmpty()) {
-            appendLine("        // No mutable non-PK columns — applyRow is a no-op.")
-        } else {
-            for (col in mutableNonPk) {
-                val rowAccess = buildRowAccess(col)
-                appendLine("        entity.${col.propertyName} = $rowAccess")
-            }
-        }
-        appendLine("    }")
-    }
-
-    private fun StringBuilder.appendApplyScalarRow(className: String, columnsIn: List<ColumnMeta>) {
-        // Override the default applyScalarRow on SqlTableDef. The default body throws — the override
-        // walks the supplied LirpRawInitializer entries, resolves each entry's Kotlin property name
-        // to its column on the table, reads the row value with the same conversion semantics as
-        // fromRow / applyRow (UUID, LocalDate, LocalDateTime, Enum), and dispatches to the entry's
-        // silentSetter so reactive backing fields are written without firing events.
-        // Skip @Embedded-derived columns: they share their top-level entity ctor-param name, which
-        // would emit duplicate `when` branches; rawInit never carries entries for embedded scalars.
-        val columns = columnsIn.filterNot { it.isInsideEmbedded }
-        appendLine("    override fun applyScalarRow(")
-        appendLine("        entity: $className,")
-        appendLine("        row: org.jetbrains.exposed.v1.core.ResultRow,")
-        appendLine("        table: org.jetbrains.exposed.v1.core.Table,")
-        appendLine("        rawInit: net.transgressoft.lirp.persistence.LirpRawInitializer<$className>")
-        appendLine("    ) {")
-        if (columns.isEmpty()) {
-            appendLine("        // No mapped columns — applyScalarRow is a no-op.")
-        } else {
-            appendLine("        for (entry in rawInit.entries) {")
-            appendLine("            val value: Any? = when (entry.name) {")
-            for (col in columns) {
-                val rowAccess = buildRowAccess(col)
-                appendLine("                \"${col.propertyName}\" -> $rowAccess")
-            }
-            appendLine("                else -> continue")
-            appendLine("            }")
-            appendLine("            entry.silentSetter(entity, value)")
-            appendLine("        }")
-        }
-        appendLine("    }")
-    }
-
-    private fun StringBuilder.appendBumpVersion(className: String, columns: List<ColumnMeta>) {
-        // Emit a non-default bumpVersion override only when the entity declares a @Version
-        // column. Unversioned entities inherit the interface no-op default, so no emission keeps
-        // the generated file minimal.
-        val versionCol = columns.singleOrNull { it.isVersion } ?: return
-        appendLine()
-        appendLine("    override fun bumpVersion(entity: $className, newVersion: Long) {")
-        appendLine("        entity.${versionCol.propertyName} = newVersion")
-        appendLine("    }")
-    }
-
     private fun resolveTableName(classDecl: KSClassDeclaration, className: String): String {
         val mappingAnnotation =
             classDecl.annotations.firstOrNull {
@@ -941,1181 +482,4 @@ class TableDefProcessor(
         val customName = mappingAnnotation?.arguments?.firstOrNull { it.name?.asString() == "name" }?.value as? String
         return if (!customName.isNullOrEmpty()) customName else className.toSnakeCase()
     }
-
-    /**
-     * Output of [collectColumnsAndSlots] — the flat column list (what `columns`, `applyRow`,
-     * `applyScalarRow`, `toParams` consume) plus the structured ctor-slot tree (what `fromRow`
-     * consumes to emit nested constructor expressions for `@Embedded` parameters).
-     */
-    private data class CollectedShape(
-        val columns: List<ColumnMeta>,
-        val ctorSlots: List<CtorSlot>
-    )
-
-    /**
-     * Walks the entity's primary constructor in declaration order, routing each parameter to one
-     * of three paths: an `@Embedded` parameter triggers recursive descent into the referenced
-     * `@Embeddable` (flattening its scalars and recording an [EmbeddedCtorSlot]); a regular
-     * parameter goes through [buildColumnMeta] and yields a [ScalarCtorSlot]; non-ctor properties
-     * (setter cols) are scanned afterwards. The result preserves ctor-arg order so the generated
-     * `fromRow` can emit positional constructor calls.
-     */
-    private fun collectColumnsAndSlots(
-        classDecl: KSClassDeclaration,
-        versionedProperty: KSPropertyDeclaration?,
-        excludedBackingFields: Set<String> = emptySet(),
-        aggregateBackingScalarNames: Set<String> = emptySet()
-    ): CollectedShape {
-        val hasDeclaredId = classDecl.getDeclaredProperties().any { it.simpleName.asString() == "id" && !it.isAbstract() }
-        val versionedName = versionedProperty?.simpleName?.asString()
-        val ctorParams = classDecl.primaryConstructor?.parameters.orEmpty()
-        val ctorParamNames = ctorParams.mapNotNull { it.name?.asString() }.toSet()
-        val propertiesByName = classDecl.getAllProperties().associateBy { it.simpleName.asString() }
-
-        val columns = mutableListOf<ColumnMeta>()
-        val ctorSlots = mutableListOf<CtorSlot>()
-
-        // D-06 (body-declared): @Embedded is constructor-only. Any non-ctor property carrying
-        // @Embedded is rejected with a single diagnostic per occurrence; the property is then
-        // ignored by the rest of the pipeline (the non-ctor scan below filters it out via
-        // `isExcluded()` checks already in place, but @Embedded body-declared props would
-        // otherwise silently fall through buildColumnMeta as unsupported types).
-        for (prop in classDecl.getAllProperties()) {
-            val propName = prop.simpleName.asString()
-            if (propName in ctorParamNames) continue
-            val hasEmbedded =
-                prop.annotations.any {
-                    it.annotationType.resolve().declaration.qualifiedName?.asString() == EMBEDDED_FQN
-                }
-            if (!hasEmbedded) continue
-            logger.error(
-                "@Embedded must be on a primary-constructor parameter (found body-declared property): " +
-                    "${classDecl.qualifiedName?.asString() ?: classDecl.simpleName.asString()}.$propName",
-                prop
-            )
-        }
-
-        for (param in ctorParams) {
-            val paramName = param.name?.asString() ?: continue
-            if (paramName in excludedBackingFields) continue
-            val prop = propertiesByName[paramName] ?: continue
-            if (prop.isExcluded()) continue
-
-            val embeddedAnnotation =
-                prop.annotations.firstOrNull {
-                    it.annotationType.resolve().declaration.qualifiedName?.asString() == EMBEDDED_FQN
-                }
-            if (embeddedAnnotation != null) {
-                if (!validateEmbeddedTargetStrictness(classDecl, param, prop)) continue
-                val slot =
-                    buildEmbeddedSlot(
-                        prop = prop,
-                        ctorParamName = paramName,
-                        embeddedAnnotation = embeddedAnnotation,
-                        columnsAccumulator = columns,
-                        parentPrefix = autoDerivedPrefix(paramName),
-                        parentPath = paramName,
-                        topLevelPropertyName = paramName
-                    )
-                if (slot != null) ctorSlots += slot
-            } else {
-                val col = buildColumnMeta(prop, hasDeclaredId, versionedName, ctorParamNames, aggregateBackingScalarNames) ?: continue
-                columns += col
-                ctorSlots += ScalarCtorSlot(paramName, col)
-            }
-        }
-
-        // Non-ctor scalar properties (setter cols). @Embedded is ctor-only by design (D-06 will
-        // diagnose body-level placement in plan 57-03); treat them as regular columns here.
-        for (prop in classDecl.getAllProperties()) {
-            val propName = prop.simpleName.asString()
-            if (propName in ctorParamNames) continue
-            if (prop.isExcluded() || propName in excludedBackingFields) continue
-            val col = buildColumnMeta(prop, hasDeclaredId, versionedName, ctorParamNames, aggregateBackingScalarNames) ?: continue
-            columns += col
-        }
-        return CollectedShape(columns, ctorSlots)
-    }
-
-    private fun autoDerivedPrefix(propertyName: String): String = "${propertyName.toSnakeCase()}_"
-
-    /**
-     * D-02 / D-03: after the recursive @Embedded flatten, the fully accumulated `ColumnMeta`
-     * list is grouped by SQL column name. Any duplicate-name group emits a single
-     * `logger.error()` naming every colliding property's entity-rooted access path so the
-     * diagnostic surfaces every angle of the collision (e.g. `album.performer.name` vs
-     * `label.name`). Paths are sorted before printing so the message is deterministic and test
-     * assertions remain stable.
-     *
-     * Returns `true` when at least one collision was reported (the caller must suppress codegen
-     * for the entity so consumers do not see partially-formed `_LirpTableDef` source).
-     */
-    private fun detectColumnCollisions(
-        classDecl: KSClassDeclaration,
-        columns: List<ColumnMeta>
-    ): Boolean {
-        val byColumnName = columns.groupBy { it.name }
-        val classFqn = classDecl.qualifiedName?.asString() ?: classDecl.simpleName.asString()
-        var collided = false
-        for ((columnName, group) in byColumnName) {
-            if (group.size <= 1) continue
-            val paths = group.map { "$classFqn.${it.embeddedPath}" }.sorted()
-            logger.error(
-                "Column name collision: '$columnName' is produced by multiple properties: " +
-                    paths.joinToString(", "),
-                classDecl
-            )
-            collided = true
-        }
-        return collided
-    }
-
-    /**
-     * Applies D-06 target strictness to a ctor-param `@Embedded` site. Rejects `var` constructor
-     * parameters and properties with custom getters. Body-declared placement is caught earlier in
-     * [collectColumnsAndSlots]. Returns `true` when the site is valid; emits one `logger.error()`
-     * per violation and returns `false` otherwise so the recursive descent skips the malformed slot.
-     */
-    private fun validateEmbeddedTargetStrictness(
-        ownerClass: KSClassDeclaration,
-        param: KSValueParameter,
-        prop: KSPropertyDeclaration
-    ): Boolean {
-        val propertyFqn =
-            "${ownerClass.qualifiedName?.asString() ?: ownerClass.simpleName.asString()}.${prop.simpleName.asString()}"
-        var ok = true
-        if (param.isVar) {
-            logger.error(
-                "@Embedded must be on a val constructor parameter (found var): $propertyFqn",
-                prop
-            )
-            ok = false
-        }
-        // KSP synthesizes a getter for every property (including data-class ctor `val`s); a
-        // user-authored custom getter is distinguished by its declaration origin being one of the
-        // source-language origins rather than SYNTHETIC. Filter on Origin so we only reject
-        // explicitly-declared getters.
-        val getter = prop.getter
-        if (getter != null && getter.origin != Origin.SYNTHETIC) {
-            logger.error(
-                "@Embedded property must not have a custom getter: $propertyFqn",
-                prop
-            )
-            ok = false
-        }
-        return ok
-    }
-
-    /**
-     * Applies the D-07 trio of `@Embeddable` kind diagnostics to an `@Embedded` consuming property.
-     * Returns the referenced [KSClassDeclaration] when all checks pass, `null` (after emitting a
-     * single `logger.error()`) otherwise. Validation order matches §specifics: kind checks fire
-     * before the recursive descent visits child parameters.
-     */
-    private fun validateEmbeddableTarget(prop: KSPropertyDeclaration): KSClassDeclaration? {
-        val propertyFqn =
-            "${prop.parentDeclaration?.qualifiedName?.asString() ?: ""}.${prop.simpleName.asString()}".trimStart('.')
-        val resolved = prop.type.resolve()
-        val declaration = resolved.declaration
-
-        // D-07.2: target must be a class declaration (not interface/type-alias/type-parameter).
-        val classDecl = declaration as? KSClassDeclaration
-        if (classDecl == null) {
-            val symbolName = declaration.qualifiedName?.asString() ?: declaration.simpleName.asString()
-            logger.error(
-                "@Embedded property must reference a class type: $propertyFqn references $symbolName",
-                prop
-            )
-            return null
-        }
-
-        // D-07.1: the referenced class must carry @Embeddable.
-        val hasEmbeddable =
-            classDecl.annotations.any {
-                it.annotationType.resolve().declaration.qualifiedName?.asString() == EMBEDDABLE_FQN
-            }
-        if (!hasEmbeddable) {
-            val referencedFqn = classDecl.qualifiedName?.asString() ?: classDecl.simpleName.asString()
-            logger.error(
-                "@Embedded property must reference an @Embeddable type: $propertyFqn references $referencedFqn",
-                prop
-            )
-            return null
-        }
-
-        // D-07.3: @Embeddable must be a concrete data class with a non-empty primary constructor.
-        // Single unified diagnostic covering non-class kinds, abstract/sealed, non-data, and
-        // no-primary-ctor variants.
-        val isConcreteDataClass =
-            classDecl.classKind == ClassKind.CLASS &&
-                Modifier.DATA in classDecl.modifiers &&
-                Modifier.ABSTRACT !in classDecl.modifiers &&
-                Modifier.SEALED !in classDecl.modifiers &&
-                (classDecl.primaryConstructor?.parameters?.isNotEmpty() == true)
-        if (!isConcreteDataClass) {
-            val classFqn = classDecl.qualifiedName?.asString() ?: classDecl.simpleName.asString()
-            logger.error(
-                "@Embeddable must be a concrete data class: $classFqn",
-                classDecl
-            )
-            return null
-        }
-        return classDecl
-    }
-
-    /**
-     * Recursively expands an `@Embedded` parameter into one [ColumnMeta] per scalar leaf (appended
-     * to [columnsAccumulator]) plus a parallel [EmbeddedCtorSlot] that the generated `fromRow` uses
-     * to emit a nested constructor expression. Prefixes concatenate top-down: parent prefix is
-     * prepended, child prefix is appended (`album_performer_name`). Composition with
-     * `@PersistenceProperty(converter = …)` at scalar leaves works because each leaf goes through
-     * the standard [buildColumnMeta] pipeline (Phase 56 D-05/D-08 hint refinement applies).
-     *
-     * Returns `null` when the referenced type is not a class or lacks a qualified name. Stricter
-     * validation (non-`@Embeddable` target, non-data-class declarations, `var` on the consuming
-     * property) is deferred to plan 57-03; well-formed input is assumed here.
-     */
-    private fun buildEmbeddedSlot(
-        prop: KSPropertyDeclaration,
-        ctorParamName: String,
-        embeddedAnnotation: KSAnnotation,
-        columnsAccumulator: MutableList<ColumnMeta>,
-        parentPrefix: String,
-        parentPath: String,
-        topLevelPropertyName: String
-    ): EmbeddedCtorSlot? {
-        // D-07 kind checks run BEFORE descent so the recursion sees only well-formed embeddables.
-        val typeDecl = validateEmbeddableTarget(prop) ?: return null
-        val typeFqn = typeDecl.qualifiedName?.asString() ?: return null
-
-        val explicitPrefix = embeddedAnnotation.arguments.firstOrNull { it.name?.asString() == "prefix" }?.value as? String
-        // D-05: empty prefix reverts to auto-derive. D-04: explicit non-empty prefix is opaque
-        // (appended verbatim, no shape validation).
-        val effectivePrefix = if (explicitPrefix.isNullOrEmpty()) parentPrefix else explicitPrefix
-
-        // D-06 body-declared check inside @Embeddable: any non-ctor property carrying @Embedded
-        // on the embeddable itself is rejected here.
-        val childCtorNames =
-            typeDecl.primaryConstructor?.parameters.orEmpty().mapNotNull { it.name?.asString() }.toSet()
-        for (childProp in typeDecl.getAllProperties()) {
-            val childName = childProp.simpleName.asString()
-            if (childName in childCtorNames) continue
-            val hasEmbedded =
-                childProp.annotations.any {
-                    it.annotationType.resolve().declaration.qualifiedName?.asString() == EMBEDDED_FQN
-                }
-            if (!hasEmbedded) continue
-            logger.error(
-                "@Embedded must be on a primary-constructor parameter (found body-declared property): " +
-                    "$typeFqn.$childName",
-                childProp
-            )
-        }
-
-        val childSlots = mutableListOf<CtorSlot>()
-        for (childParam in typeDecl.primaryConstructor?.parameters.orEmpty()) {
-            val childParamName = childParam.name?.asString() ?: continue
-            val childProp =
-                typeDecl.getDeclaredProperties().firstOrNull { it.simpleName.asString() == childParamName } ?: continue
-
-            val childEmbedded =
-                childProp.annotations.firstOrNull {
-                    it.annotationType.resolve().declaration.qualifiedName?.asString() == EMBEDDED_FQN
-                }
-            if (childEmbedded != null) {
-                if (!validateEmbeddedTargetStrictness(typeDecl, childParam, childProp)) continue
-                val nestedAutoDerived = "${effectivePrefix}${autoDerivedPrefix(childParamName)}"
-                val nested =
-                    buildEmbeddedSlot(
-                        prop = childProp,
-                        ctorParamName = childParamName,
-                        embeddedAnnotation = childEmbedded,
-                        columnsAccumulator = columnsAccumulator,
-                        parentPrefix = nestedAutoDerived,
-                        parentPath = "$parentPath.$childParamName",
-                        topLevelPropertyName = topLevelPropertyName
-                    ) ?: continue
-                childSlots += nested
-            } else {
-                val leafCol =
-                    buildEmbeddedLeafColumn(
-                        childProp = childProp,
-                        childParamName = childParamName,
-                        prefix = effectivePrefix,
-                        parentPath = parentPath,
-                        topLevelPropertyName = topLevelPropertyName
-                    ) ?: continue
-                columnsAccumulator += leafCol
-                childSlots += ScalarCtorSlot(childParamName, leafCol)
-            }
-        }
-        return EmbeddedCtorSlot(ctorParamName, typeFqn, childSlots)
-    }
-
-    /**
-     * Resolves the [ColumnMeta] for a single scalar leaf inside an `@Embeddable`. Routes through
-     * the same type/converter resolution path as [buildColumnMeta] — converter routing reuses
-     * Phase 56's [refineConverterSqlType] and [mapToColumnTypeExpression] — but stamps the column
-     * with the concatenated `${prefix}${snake(leaf)}` name, the embedded access path, and marks it
-     * `isInsideEmbedded = true` so downstream emitters (`applyRow`, `applyScalarRow`) skip it.
-     */
-    private fun buildEmbeddedLeafColumn(
-        childProp: KSPropertyDeclaration,
-        childParamName: String,
-        prefix: String,
-        parentPath: String,
-        topLevelPropertyName: String
-    ): ColumnMeta? {
-        val persistenceAnnotation =
-            childProp.annotations.firstOrNull {
-                it.annotationType.resolve().declaration.qualifiedName?.asString() == PERSISTENCE_PROPERTY_FQN
-            }
-        val propertyFqn = "${childProp.parentDeclaration?.qualifiedName?.asString() ?: ""}.$childParamName".trimStart('.')
-
-        val resolvedType = childProp.type.resolve()
-        val notNullable = resolvedType.makeNotNullable()
-        val childTypeFqn = notNullable.declaration.qualifiedName?.asString() ?: "kotlin.Any"
-        val isEnum = (notNullable.declaration as? KSClassDeclaration)?.classKind == ClassKind.ENUM_CLASS
-
-        // Reuse the Phase 56 converter resolution + hint-refinement pipeline so an
-        // `@PersistenceProperty(converter = X::class)` at a scalar leaf inside an `@Embeddable`
-        // produces the same column-type expression and fromRow/toParams casts it would at the
-        // top level (D-08).
-        val converterInfo =
-            if (hasNonSentinelConverterArgument(persistenceAnnotation)) {
-                resolveConverter(persistenceAnnotation, propertyFqn)
-            } else {
-                null
-            }
-
-        val hasExplicitConverter = hasNonSentinelConverterArgument(persistenceAnnotation)
-        val typeExpression =
-            if (converterInfo != null) {
-                val hints =
-                    PersistencePropertyHints(
-                        length = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "length" }?.value as? Int ?: -1,
-                        precision = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "precision" }?.value as? Int ?: -1,
-                        scale = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "scale" }?.value as? Int ?: -1,
-                        typeHint = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "type" }?.value as? String ?: ""
-                    )
-                refineConverterSqlType(converterInfo, hints, propertyFqn, childParamName) ?: return null
-            } else if (hasExplicitConverter) {
-                return null
-            } else {
-                mapToColumnTypeExpression(childProp, persistenceAnnotation) ?: return null
-            }
-
-        val columnName = "$prefix${childParamName.toSnakeCase()}"
-        return ColumnMeta(
-            name = columnName,
-            // propertyName carries the top-level entity ctor-param so the mutability gate (which
-            // exempts ctor-param val fields) recognises this column as ctor-driven. The actual
-            // entity-access path lives in embeddedPath.
-            propertyName = topLevelPropertyName,
-            typeExpression = typeExpression,
-            typeFqn = childTypeFqn,
-            nullable = resolvedType.isMarkedNullable,
-            isPrimaryKey = false,
-            isEnum = isEnum,
-            isMutable = false,
-            isCtorParam = true,
-            isVersion = false,
-            converterFqn = converterInfo?.converterFqn,
-            converterSqlFqn = converterInfo?.sqlTypeFqn,
-            embeddedPath = "$parentPath.$childParamName",
-            isInsideEmbedded = true
-        )
-    }
-
-    private fun buildColumnMeta(
-        prop: KSPropertyDeclaration,
-        hasDeclaredId: Boolean,
-        versionedName: String?,
-        ctorParamNames: Set<String>,
-        aggregateBackingScalarNames: Set<String> = emptySet()
-    ): ColumnMeta? {
-        val propName = prop.simpleName.asString()
-        val persistenceAnnotation =
-            prop.annotations.firstOrNull {
-                it.annotationType.resolve().declaration.qualifiedName?.asString() == PERSISTENCE_PROPERTY_FQN
-            }
-        val resolvedType = prop.type.resolve()
-        val notNullableType = resolvedType.makeNotNullable()
-        val typeFqn = notNullableType.declaration.qualifiedName?.asString() ?: "kotlin.Any"
-        val isEnum = (notNullableType.declaration as? KSClassDeclaration)?.classKind == ClassKind.ENUM_CLASS
-
-        val propertyFqn = "${prop.parentDeclaration?.qualifiedName?.asString() ?: ""}.$propName".trimStart('.')
-
-        val isPrimaryKey = propName == "id" && hasDeclaredId && !prop.isAbstract()
-        val isVersion = versionedName != null && propName == versionedName
-        val isAggregateBackingScalar = propName in aggregateBackingScalarNames
-
-        // D-06: reject converter arguments on PK / @Version / @Aggregate single-ref FK columns
-        // BEFORE invoking resolveConverter (which carries D-07 / D-08). This preserves the
-        // one-diagnostic-per-site invariant: a misplaced converter on a rejected target emits
-        // the target rejection, not the kind/S diagnostics.
-        val converterInfo =
-            if (hasNonSentinelConverterArgument(persistenceAnnotation)) {
-                when {
-                    isPrimaryKey -> {
-                        logger.error(
-                            "@PersistenceProperty(converter = …) is not allowed on primary key column '$propertyFqn'. " +
-                                "Converters apply only to non-PK scalar columns; domain-typed identifiers are deferred to a future phase."
-                        )
-                        null
-                    }
-                    isVersion -> {
-                        logger.error(
-                            "@PersistenceProperty(converter = …) is not allowed on @Version column '$propertyFqn'. " +
-                                "@Version columns require a numeric (Long) scalar type; converter routing on optimistic-locking columns is out of scope."
-                        )
-                        null
-                    }
-                    isAggregateBackingScalar -> {
-                        logger.error(
-                            "@PersistenceProperty(converter = …) is not allowed on @Aggregate single-ref FK scalar column '$propertyFqn'. " +
-                                "Converter routing on aggregate FK columns is out of scope; the FK column type is dictated by the referenced entity's primary key type."
-                        )
-                        null
-                    }
-                    else -> resolveConverter(persistenceAnnotation, propertyFqn)
-                }
-            } else {
-                null
-            }
-
-        // When a converter is bound, derive the column type from the converter's declared sqlType
-        // (and refine it via compatible @PersistenceProperty hints). Otherwise fall back to the
-        // FQN-driven base expression. A null refinement result means a hint/converter mismatch was
-        // diagnosed via logger.error — drop the column so codegen does not emit a malformed file.
-        // The non-converter branch invokes mapToColumnTypeExpression lazily so a converter-bound
-        // column with a non-scalar Kotlin type (e.g. a value class) bypasses the
-        // "unsupported column type" diagnostic that path would otherwise emit.
-        val hasExplicitConverter = hasNonSentinelConverterArgument(persistenceAnnotation)
-        val typeExpression =
-            if (converterInfo != null) {
-                val hints =
-                    PersistencePropertyHints(
-                        length = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "length" }?.value as? Int ?: -1,
-                        precision = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "precision" }?.value as? Int ?: -1,
-                        scale = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "scale" }?.value as? Int ?: -1,
-                        typeHint = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "type" }?.value as? String ?: ""
-                    )
-                refineConverterSqlType(
-                    converterInfo = converterInfo, hints = hints, propertyFqn = propertyFqn, propName = propName
-                ) ?: return null
-            } else if (hasExplicitConverter) {
-                // Converter declared but unresolved/invalid this round; avoid non-converter fallback.
-                // Invalid converters already emitted diagnostics in resolveConverter.
-                return null
-            } else {
-                mapToColumnTypeExpression(prop, persistenceAnnotation) ?: return null
-            }
-
-        return ColumnMeta(
-            name = columnNameFor(persistenceAnnotation, propName),
-            propertyName = propName,
-            typeExpression = typeExpression,
-            typeFqn = typeFqn,
-            nullable = resolvedType.isMarkedNullable,
-            isPrimaryKey = isPrimaryKey,
-            isEnum = isEnum,
-            isMutable = prop.isMutable && hasPublicSetter(prop),
-            isCtorParam = propName in ctorParamNames,
-            isVersion = isVersion,
-            converterFqn = converterInfo?.converterFqn,
-            converterSqlFqn = converterInfo?.sqlTypeFqn
-        )
-    }
-
-    /**
-     * Resolves the generated `typeExpression` for a converter-bearing column.
-     *
-     * The converter's declared `sqlType` (resolved at KSP time via `converterSqlFqn`) is the base.
-     * `@PersistenceProperty` hints layer on top:
-     *  - non-empty `type=...` always wins (explicit consumer intent — delegates to the shared
-     *    [mapTypeHintToExpression] helper used by non-converter columns);
-     *  - `length` on a `TextType` base refines to `VarcharType(length)`;
-     *  - `precision`/`scale` on a numeric base (Int/Short/Byte/Long/Double/Float/BigDecimal) refines
-     *    to `DecimalType(precision, scale)` — overrides the converter's declared precision/scale
-     *    when both are present;
-     *  - any other base + any hint set is rejected with a KSP error naming the property, the
-     *    converter, and the resolved sqlType.
-     *
-     * Returns null when an incompatible hint was diagnosed; in that case the caller drops the
-     * column so codegen does not emit a malformed file.
-     */
-    private fun refineConverterSqlType(
-        converterInfo: ConverterInfo,
-        hints: PersistencePropertyHints,
-        propertyFqn: String,
-        propName: String
-    ): String? {
-        val converterFqn = converterInfo.converterFqn
-        val converterSqlFqn = converterInfo.sqlTypeFqn
-        val length = hints.length
-        val precision = hints.precision
-        val scale = hints.scale
-        // Non-empty `type = "..."` is explicit consumer intent — delegate to the existing hint
-        // mapper so converter columns honor the same vocabulary (TEXT/VARCHAR/INT/...) as
-        // non-converter columns.
-        if (hints.typeHint.isNotEmpty()) {
-            return mapTypeHintToExpression(hints.typeHint, length, precision, scale, propName)
-        }
-
-        val baseExpression = "$converterFqn.sqlType"
-        val hasLength = length > 0
-        val hasPrecisionOrScale = precision > 0 || scale >= 0
-        if (!hasLength && !hasPrecisionOrScale) return baseExpression
-
-        val resolvedBase = SUPPORTED_CONVERTER_S_TYPES[converterSqlFqn]
-        return when (converterSqlFqn) {
-            KOTLIN_STRING_FQN -> {
-                if (hasPrecisionOrScale) {
-                    logger.error(
-                        "@PersistenceProperty hint precision/scale on property '$propertyFqn' is " +
-                            "incompatible with converter '$converterFqn' whose sqlType resolves to " +
-                            "$resolvedBase. Remove the hint or pick a converter with a numeric sqlType."
-                    )
-                    return null
-                }
-                "ColumnType.VarcharType($length)"
-            }
-            KOTLIN_INT_FQN, KOTLIN_SHORT_FQN, KOTLIN_BYTE_FQN, KOTLIN_LONG_FQN,
-            KOTLIN_DOUBLE_FQN, KOTLIN_FLOAT_FQN, BIG_DECIMAL_FQN -> {
-                if (hasLength) {
-                    logger.error(
-                        "@PersistenceProperty hint length on property '$propertyFqn' is " +
-                            "incompatible with converter '$converterFqn' whose sqlType resolves to " +
-                            "$resolvedBase. Remove the hint or pick a converter with a textual sqlType."
-                    )
-                    return null
-                }
-                val p = if (precision > 0) precision else 19
-                val s = if (scale >= 0) scale else 2
-                "ColumnType.DecimalType($p, $s)"
-            }
-            else -> {
-                logger.error(
-                    "@PersistenceProperty hints (length/precision/scale) on property '$propertyFqn' are " +
-                        "incompatible with converter '$converterFqn' whose sqlType resolves to " +
-                        "$resolvedBase. Remove the hint or pick a converter with a compatible sqlType."
-                )
-                null
-            }
-        }
-    }
-
-    /**
-     * Tests whether the `converter` argument of `@PersistenceProperty` resolves to a non-sentinel
-     * class declaration. The sentinel [ColumnConverter][net.transgressoft.lirp.persistence.ColumnConverter]
-     * interface FQN means "no converter declared"; any other class triggers D-06 / D-07 / D-08
-     * validation downstream.
-     */
-    private fun hasNonSentinelConverterArgument(annotation: KSAnnotation?): Boolean {
-        if (annotation == null) return false
-        val converterArg = annotation.arguments.firstOrNull { it.name?.asString() == "converter" }?.value as? KSType ?: return false
-        val converterFqn = converterArg.declaration.qualifiedName?.asString() ?: return false
-        return converterFqn != COLUMN_CONVERTER_FQN
-    }
-
-    /**
-     * Reads the `converter` argument from a `@PersistenceProperty` annotation and validates
-     * the referenced [ColumnConverter][net.transgressoft.lirp.persistence.ColumnConverter]
-     * singleton against the structural contract (D-07 object kind, D-08 supported S type).
-     *
-     * Returns null when no converter is declared (sentinel), when the converter cannot yet
-     * be resolved in this round (validate() guard so cross-round resolution does not produce
-     * spurious diagnostics), or when validation fails after a diagnostic has been logged.
-     */
-    private fun resolveConverter(annotation: KSAnnotation?, propertyFqn: String): ConverterInfo? {
-        if (annotation == null) return null
-        val converterArg = annotation.arguments.firstOrNull { it.name?.asString() == "converter" }?.value as? KSType ?: return null
-        val converterDecl = converterArg.declaration as? KSClassDeclaration ?: return null
-        val converterFqn = converterDecl.qualifiedName?.asString() ?: return null
-
-        // Sentinel: the interface itself means "no converter declared" — silent skip.
-        if (converterFqn == COLUMN_CONVERTER_FQN) return null
-
-        // Defer to a later KSP round when the converter type cannot yet be fully resolved.
-        if (!converterDecl.validate()) return null
-
-        if (converterDecl.classKind != ClassKind.OBJECT) {
-            logger.error(
-                "Converter '$converterFqn' for property '$propertyFqn' must be a Kotlin `object` " +
-                    "(singleton) so KSP-generated code can reference it without instantiation."
-            )
-            return null
-        }
-
-        val columnConverterSuperType =
-            converterDecl.superTypes
-                .map { it.resolve() }
-                .firstOrNull { it.declaration.qualifiedName?.asString() == COLUMN_CONVERTER_FQN }
-
-        // S is the SECOND type argument of ColumnConverter<D, S> (index 1, not 0).
-        val sTypeArg = columnConverterSuperType?.arguments?.getOrNull(1)?.type?.resolve()
-        val sFqn = sTypeArg?.declaration?.qualifiedName?.asString()
-
-        if (columnConverterSuperType == null || sFqn == null) {
-            logger.error(
-                "Converter '$converterFqn' does not declare ColumnConverter<D, S> as a supertype " +
-                    "with both type arguments resolved."
-            )
-            return null
-        }
-
-        if (sFqn !in SUPPORTED_CONVERTER_S_TYPES) {
-            logger.error(
-                "Converter '$converterFqn' declares S=$sFqn which is not supported. Supported types: " +
-                    "kotlin.String, kotlin.Int, kotlin.Long, kotlin.Short, kotlin.Byte, kotlin.Boolean, " +
-                    "kotlin.Double, kotlin.Float, java.math.BigDecimal, java.util.UUID, " +
-                    "java.time.LocalDate, java.time.LocalDateTime."
-            )
-            return null
-        }
-
-        return ConverterInfo(converterFqn = converterFqn, sqlTypeFqn = sFqn)
-    }
-
-    private fun columnNameFor(persistenceAnnotation: KSAnnotation?, propName: String): String {
-        if (persistenceAnnotation == null) return propName.toSnakeCase()
-        val customName = persistenceAnnotation.arguments.firstOrNull { it.name?.asString() == "name" }?.value as? String
-        return if (!customName.isNullOrEmpty()) customName else propName.toSnakeCase()
-    }
-
-    // Mutable for SqlTableDef fromRow purposes means: var property with a public setter.
-    private fun hasPublicSetter(prop: KSPropertyDeclaration): Boolean =
-        prop.setter?.modifiers?.none {
-            it == Modifier.PROTECTED || it == Modifier.PRIVATE || it == Modifier.INTERNAL
-        } ?: true
-
-    private fun KSPropertyDeclaration.isExcluded(): Boolean {
-        // Private backing fields are encapsulated implementation state — no public surface for
-        // persistence to bind through. Mirrors the exclusion applied by `RawInitializerProcessor`
-        // so the two processors agree on the persisted column set for a given entity shape.
-        if (Modifier.PRIVATE in modifiers) return true
-        val annotationFqns =
-            annotations
-                .map { it.annotationType.resolve().declaration.qualifiedName?.asString() }
-                .toSet()
-        if (PERSISTENCE_IGNORE_FQN in annotationFqns) return true
-        if (AGGREGATE_ANNOTATION_FQN in annotationFqns) return true
-        if (TRANSIENT_FQN in annotationFqns) return true
-        // Exclude computed properties (no backing field, not delegated), but include delegate-backed properties
-        if (!hasBackingField && !isDelegated()) return true
-        return false
-    }
-
-    private fun mapToColumnTypeExpression(
-        prop: KSPropertyDeclaration,
-        persistenceAnnotation: KSAnnotation?
-    ): String? {
-        val resolvedType = prop.type.resolve()
-        val notNullableType = resolvedType.makeNotNullable()
-        val fqn = notNullableType.declaration.qualifiedName?.asString()
-
-        val length = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "length" }?.value as? Int ?: -1
-        val precision = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "precision" }?.value as? Int ?: -1
-        val scale = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "scale" }?.value as? Int ?: -1
-        val typeHint = persistenceAnnotation?.arguments?.firstOrNull { it.name?.asString() == "type" }?.value as? String ?: ""
-
-        // Explicit type hint takes precedence over FQN-based inference
-        if (typeHint.isNotEmpty()) {
-            return mapTypeHintToExpression(typeHint, length, precision, scale, prop.simpleName.asString())
-        }
-
-        // Short and Byte map to IntType (no dedicated ColumnType variant). SQLite stores all
-        // integer affinities as INTEGER regardless of declared width, so introducing SMALLINT /
-        // TINYINT variants would force dialect-specific SQL generation for marginal benefit.
-        // The Kotlin side narrows on read (`Int.toShort()` / `Int.toByte()`) and widens on write
-        // (`Short.toInt()` / `Byte.toInt()`) — see buildRowAccess and buildEntityAccess below.
-        return when (fqn) {
-            KOTLIN_INT_FQN -> COLUMN_TYPE_INT_EXPR
-            KOTLIN_SHORT_FQN -> COLUMN_TYPE_INT_EXPR
-            KOTLIN_BYTE_FQN -> COLUMN_TYPE_INT_EXPR
-            KOTLIN_LONG_FQN -> COLUMN_TYPE_LONG_EXPR
-            KOTLIN_STRING_FQN -> if (length > 0) "ColumnType.VarcharType($length)" else COLUMN_TYPE_TEXT_EXPR
-            KOTLIN_BOOLEAN_FQN -> COLUMN_TYPE_BOOLEAN_EXPR
-            KOTLIN_DOUBLE_FQN -> COLUMN_TYPE_DOUBLE_EXPR
-            KOTLIN_FLOAT_FQN -> COLUMN_TYPE_FLOAT_EXPR
-            UUID_FQN -> COLUMN_TYPE_UUID_EXPR
-            LOCAL_DATE_TIME_FQN -> COLUMN_TYPE_DATETIME_EXPR
-            LOCAL_DATE_FQN -> COLUMN_TYPE_DATE_EXPR
-            BIG_DECIMAL_FQN -> {
-                val p = if (precision > 0) precision else 19
-                val s = if (scale >= 0) scale else 2
-                "ColumnType.DecimalType($p, $s)"
-            }
-            else -> {
-                val declaration = notNullableType.declaration
-                if ((declaration as? KSClassDeclaration)?.classKind == ClassKind.ENUM_CLASS) {
-                    "ColumnType.EnumType(\"$fqn\")"
-                } else {
-                    logger.error("Unsupported column type '$fqn' on property '${prop.simpleName.asString()}' — cannot map to ColumnType")
-                    null
-                }
-            }
-        }
-    }
-
-    private fun mapTypeHintToExpression(
-        hint: String,
-        length: Int,
-        precision: Int,
-        scale: Int,
-        propName: String
-    ): String? =
-        when (hint.uppercase()) {
-            "TEXT" -> COLUMN_TYPE_TEXT_EXPR
-            "VARCHAR" -> {
-                if (length <= 0) {
-                    logger.error("@PersistenceProperty(type=\"VARCHAR\") requires length > 0 on property '$propName'")
-                    return null
-                }
-                "ColumnType.VarcharType($length)"
-            }
-            "INT" -> COLUMN_TYPE_INT_EXPR
-            "BIGINT" -> COLUMN_TYPE_LONG_EXPR
-            "BOOLEAN" -> COLUMN_TYPE_BOOLEAN_EXPR
-            "DOUBLE" -> COLUMN_TYPE_DOUBLE_EXPR
-            "FLOAT" -> COLUMN_TYPE_FLOAT_EXPR
-            "UUID" -> COLUMN_TYPE_UUID_EXPR
-            "DATE" -> COLUMN_TYPE_DATE_EXPR
-            "DATETIME" -> COLUMN_TYPE_DATETIME_EXPR
-            "DECIMAL" -> {
-                val p = if (precision > 0) precision else 19
-                val s = if (scale >= 0) scale else 2
-                "ColumnType.DecimalType($p, $s)"
-            }
-            else -> {
-                logger.error("Unknown @PersistenceProperty type hint '$hint' on property '$propName'")
-                null
-            }
-        }
-
-    /**
-     * Builds the [ForeignKeyMeta] list for an entity by walking its single-entity `@Aggregate`
-     * properties. Collection refs are skipped — they are handled by junction-table descriptors.
-     *
-     * Validates each single-entity ref:
-     *  - The lambda body of `aggregate { … }` must be a bare identifier naming the backing scalar.
-     *  - The backing scalar property must exist on the same class.
-     *  - `@Aggregate(onDelete = DETACH)` requires the backing scalar to be nullable (Spike 006).
-     *
-     * Drops entries with `onDelete = NONE` — by convention, NONE means "no FK clause at all".
-     */
-    private fun collectForeignKeys(
-        classDecl: KSClassDeclaration,
-        aggregates: List<AggregatePropertyMeta>
-    ): List<ForeignKeyMeta> {
-        if (aggregates.isEmpty()) return emptyList()
-
-        val foreignKeys = mutableListOf<ForeignKeyMeta>()
-        val propertiesByName = classDecl.getAllProperties().associateBy { it.simpleName.asString() }
-
-        for (agg in aggregates.filter { !it.isCollection }) {
-            val propName = agg.propertyName
-            val scalarName =
-                agg.backingScalarName
-                    ?: run {
-                        logger.error(
-                            "Cannot determine backing scalar for @Aggregate property '$propName'. " +
-                                "The aggregate { … } lambda must reference exactly one scalar property.",
-                            agg.property
-                        )
-                        continue
-                    }
-
-            val scalarProp = propertiesByName[scalarName]
-            if (scalarProp == null) {
-                logger.error(
-                    "@Aggregate property '$propName' references unknown scalar '$scalarName'.",
-                    agg.property
-                )
-                continue
-            }
-
-            val onDelete = agg.onDeleteName
-            if (onDelete == "DETACH" && !scalarProp.type.resolve().isMarkedNullable) {
-                logger.error(
-                    "@Aggregate(onDelete = DETACH) on property '$propName' requires a nullable backing scalar. " +
-                        "Make '$scalarName' nullable (e.g., 'Long?') or choose a different CascadeAction " +
-                        "(RESTRICT, CASCADE, NONE).",
-                    agg.property
-                )
-                continue
-            }
-
-            // NONE => emit no FK clause at all (preserves backwards compatibility per Spike 006).
-            if (onDelete == "NONE") continue
-
-            val referencedTableName = resolveTableName(agg.referencedClass, agg.referencedClass.simpleName.asString())
-
-            // Resolve the local FK column name using @PersistenceProperty(name=...) on the backing
-            // scalar, mirroring the column-name resolution logic in collectColumns().
-            val scalarPersistenceAnnotation =
-                scalarProp.annotations.firstOrNull {
-                    it.annotationType.resolve().declaration.qualifiedName?.asString() == PERSISTENCE_PROPERTY_FQN
-                }
-            val localColumnName =
-                (
-                    scalarPersistenceAnnotation
-                        ?.arguments
-                        ?.firstOrNull { it.name?.asString() == "name" }
-                        ?.value as? String
-                )
-                    ?.takeIf { it.isNotEmpty() }
-                    ?: scalarName.toSnakeCase()
-
-            // Resolve the referenced entity's PK column name using @PersistenceProperty(name=...)
-            // on the 'id' property of the referenced entity class.
-            val referencedIdPropForFk =
-                agg.referencedClass.getAllProperties()
-                    .firstOrNull { it.simpleName.asString() == "id" }
-            val referencedPkAnnotation =
-                referencedIdPropForFk?.annotations?.firstOrNull {
-                    it.annotationType.resolve().declaration.qualifiedName?.asString() == PERSISTENCE_PROPERTY_FQN
-                }
-            val referencedColumnName =
-                (
-                    referencedPkAnnotation
-                        ?.arguments
-                        ?.firstOrNull { it.name?.asString() == "name" }
-                        ?.value as? String
-                )
-                    ?.takeIf { it.isNotEmpty() }
-                    ?: "id"
-
-            foreignKeys.add(
-                ForeignKeyMeta(
-                    columnName = localColumnName,
-                    referencedTable = referencedTableName,
-                    referencedColumn = referencedColumnName,
-                    onDelete = onDelete
-                )
-            )
-        }
-        return foreignKeys
-    }
-
-    /**
-     * Builds the [JunctionRefInfo] list for one entity by walking its collection-typed `@Aggregate`
-     * properties.
-     *
-     * Validates each collection ref:
-     *  - The first positional argument of `aggregateList(…)` / `aggregateSet(…)` must be a bare
-     *    identifier naming a property on the same class.
-     *  - That property must be `var`, with a stdlib `kotlin.collections.List`,
-     *    `kotlin.collections.MutableList`, `kotlin.collections.Set`, or
-     *    `kotlin.collections.MutableSet` type whose element type matches the item entity's `id`
-     *    type.
-     *  - For `aggregateList`, the property must be a `List` / `MutableList`. For `aggregateSet`,
-     *    it must be a `Set` / `MutableSet`.
-     *
-     * Failures emit `KSP[FK-04]` errors and skip emission for the affected entity. The successful
-     * entries drive the `junctionTableDefs` / `junctionAccessors` / `applyJunctionRows` overrides
-     * on the parent's `_LirpTableDef`.
-     */
-    private fun collectJunctionRefs(
-        classDecl: KSClassDeclaration,
-        aggregates: List<AggregatePropertyMeta>
-    ): List<JunctionRefInfo> {
-        val collectionAggs = aggregates.filter { it.isCollection }
-        if (collectionAggs.isEmpty()) return emptyList()
-
-        val propertiesByName = classDecl.getAllProperties().associateBy { it.simpleName.asString() }
-        val parentSimpleName = classDecl.simpleName.asString()
-        return collectionAggs.mapNotNull { buildJunctionRefInfo(it, parentSimpleName, propertiesByName) }
-    }
-
-    private fun buildJunctionRefInfo(
-        agg: AggregatePropertyMeta,
-        parentSimpleName: String,
-        propertiesByName: Map<String, KSPropertyDeclaration>
-    ): JunctionRefInfo? {
-        val backingName =
-            agg.backingCollectionName
-                ?: return logFk04MissingBacking(agg, parentSimpleName).let { null }
-        val backingProp =
-            propertiesByName[backingName]
-                ?: return logFk04MissingBackingProp(agg, parentSimpleName, backingName).let { null }
-        if (!backingProp.isMutable) return logFk04ImmutableBacking(agg, parentSimpleName, backingName).let { null }
-
-        val resolvedType = backingProp.type.resolve()
-        val typeFqn = resolvedType.makeNotNullable().declaration.qualifiedName?.asString()
-        if (!validateBackingShape(agg, parentSimpleName, backingName, typeFqn)) return null
-
-        val elementFqn =
-            resolvedType.arguments.firstOrNull()?.type?.resolve()
-                ?.makeNotNullable()?.declaration?.qualifiedName?.asString()
-        if (!validateElementMatchesReferencedId(agg, backingName, elementFqn)) return null
-
-        val itemSimpleName = elementFqnToSimpleName(elementFqn)
-        val propertyCapitalized = agg.propertyName.replaceFirstChar { it.uppercase() }
-        val descriptorName = "${parentSimpleName}_${propertyCapitalized}_LirpJunctionTableDef"
-        val isMutableList = typeFqn == "kotlin.collections.MutableList" || typeFqn == "kotlin.collections.MutableSet"
-
-        return JunctionRefInfo(
-            propertyName = agg.propertyName,
-            backingFieldName = backingName,
-            junctionObjectName = descriptorName,
-            isOrdered = agg.isOrdered,
-            itemKeyTypeSimpleName = itemSimpleName,
-            isMutableList = isMutableList
-        )
-    }
-
-    private fun logFk04MissingBacking(agg: AggregatePropertyMeta, parentSimpleName: String) {
-        logger.error(
-            "KSP[FK-04]: @Aggregate collection property '${agg.propertyName}' on " +
-                "'$parentSimpleName' must be a 'var List<K>'/'var Set<K>' bound to a writable " +
-                "backing field passed as the first positional argument to " +
-                "${if (agg.isOrdered) "aggregateList" else "aggregateSet"}(<field>). " +
-                "Anonymous initialisers like 'emptyList()' or 'setOf(...)' are not supported.",
-            agg.property
-        )
-    }
-
-    private fun logFk04MissingBackingProp(agg: AggregatePropertyMeta, parentSimpleName: String, backingName: String) {
-        logger.error(
-            "KSP[FK-04]: backing field '$backingName' for @Aggregate property " +
-                "'${agg.propertyName}' on '$parentSimpleName' must be a 'var List<K>'/" +
-                "'var Set<K>' declared on the same class.",
-            agg.property
-        )
-    }
-
-    private fun logFk04ImmutableBacking(agg: AggregatePropertyMeta, parentSimpleName: String, backingName: String) {
-        logger.error(
-            "KSP[FK-04]: backing field '$backingName' for @Aggregate property " +
-                "'${agg.propertyName}' on '$parentSimpleName' must be a 'var List<K>'/" +
-                "'var Set<K>' (declared 'val').",
-            agg.property
-        )
-    }
-
-    private fun validateBackingShape(
-        agg: AggregatePropertyMeta,
-        parentSimpleName: String,
-        backingName: String,
-        typeFqn: String?
-    ): Boolean {
-        val isList = typeFqn == "kotlin.collections.List" || typeFqn == "kotlin.collections.MutableList"
-        val isSet = typeFqn == "kotlin.collections.Set" || typeFqn == "kotlin.collections.MutableSet"
-        if (agg.isOrdered && !isList) {
-            logger.error(
-                "KSP[FK-04]: backing field '$backingName' for @Aggregate property " +
-                    "'${agg.propertyName}' on '$parentSimpleName' must be a 'var List<K>' for " +
-                    "aggregateList; found '$typeFqn'.",
-                agg.property
-            )
-            return false
-        }
-        if (!agg.isOrdered && !isSet) {
-            logger.error(
-                "KSP[FK-04]: backing field '$backingName' for @Aggregate property " +
-                    "'${agg.propertyName}' on '$parentSimpleName' must be a 'var Set<K>' for " +
-                    "aggregateSet; found '$typeFqn'.",
-                agg.property
-            )
-            return false
-        }
-        return true
-    }
-
-    private fun validateElementMatchesReferencedId(
-        agg: AggregatePropertyMeta,
-        backingName: String,
-        elementFqn: String?
-    ): Boolean {
-        // KSP[FK-05]: verify the backing collection's element type matches the referenced
-        // entity's ID type. A mismatch (e.g. List<Int> backing an aggregate whose target uses
-        // Long ids) is silently accepted by the compiler but causes filterIsInstance to drop
-        // all loaded IDs at runtime.
-        val referencedIdProp =
-            agg.referencedClass.getAllProperties()
-                .firstOrNull { it.simpleName.asString() == "id" } ?: return true
-        val referencedIdFqn =
-            referencedIdProp.type.resolve().makeNotNullable()
-                .declaration.qualifiedName?.asString() ?: return true
-        if (elementFqn == null || elementFqn == referencedIdFqn) return true
-
-        // Normalize kotlin.UUID alias to java.util.UUID for comparison.
-        val normalizedElement = if (elementFqn == KOTLIN_UUID_FQN) UUID_FQN else elementFqn
-        val normalizedRefId = if (referencedIdFqn == KOTLIN_UUID_FQN) UUID_FQN else referencedIdFqn
-        if (normalizedElement == normalizedRefId) return true
-
-        logger.error(
-            "KSP[FK-05]: backing field '$backingName' element type '$elementFqn' does not match " +
-                "the referenced entity '${agg.referencedClass.simpleName.asString()}' ID type " +
-                "'$referencedIdFqn'. Fix the backing collection's type parameter to match.",
-            agg.property
-        )
-        return false
-    }
-
-    private fun elementFqnToSimpleName(elementFqn: String?): String =
-        when (elementFqn) {
-            KOTLIN_INT_FQN -> "Int"
-            KOTLIN_LONG_FQN -> "Long"
-            KOTLIN_STRING_FQN -> "String"
-            KOTLIN_UUID_FQN, UUID_FQN -> "java.util.UUID"
-            null -> "Any"
-            else -> elementFqn.substringAfterLast(".")
-        }
-
-    private fun extractCascadeActionName(value: Any?): String =
-        when {
-            value is KSType -> value.declaration.simpleName.asString()
-            value != null -> {
-                val str = value.toString()
-                when {
-                    str.endsWith("CASCADE") -> "CASCADE"
-                    str.endsWith("NONE") -> "NONE"
-                    str.endsWith("RESTRICT") -> "RESTRICT"
-                    else -> "DETACH"
-                }
-            }
-            else -> "DETACH"
-        }
-
-    /**
-     * Emits a `{Parent}_{Property}_LirpJunctionTableDef` object that implements [JunctionTableDef]
-     * for one collection-typed `@Aggregate` property.
-     *
-     * The descriptor is the SQL-side companion of the parent's `_LirpTableDef` and lives in the
-     * same package. Its column shape is fixed: `(parent_id, item_id)` always form the composite
-     * primary key; `position` is appended for `aggregateList` and omitted for `aggregateSet`.
-     */
-    private fun generateJunctionTableDef(
-        parentClass: KSClassDeclaration,
-        agg: AggregatePropertyMeta
-    ) {
-        val packageName = parentClass.packageName.asString()
-        val parentSimpleName = parentClass.simpleName.asString()
-        val itemSimpleName = agg.referencedClass.simpleName.asString()
-        val propertyCapitalized = agg.propertyName.replaceFirstChar { it.uppercase() }
-        val descriptorName = "${parentSimpleName}_${propertyCapitalized}_LirpJunctionTableDef"
-
-        val parentTableName = resolveTableName(parentClass, parentSimpleName)
-        val itemTableName = resolveTableName(agg.referencedClass, itemSimpleName)
-        val junctionTableName = "${parentTableName}_${agg.propertyName.toSnakeCase()}"
-
-        val parentPkType = pkColumnTypeExpression(parentClass) ?: COLUMN_TYPE_INT_EXPR
-        val itemPkType = pkColumnTypeExpression(agg.referencedClass) ?: COLUMN_TYPE_INT_EXPR
-
-        val file =
-            codeGenerator.createNewFile(
-                dependencies = Dependencies(false, parentClass.containingFile!!),
-                packageName = packageName,
-                fileName = descriptorName
-            )
-
-        // Item-side cascade action defaults to DETACH per @Aggregate's annotation default; that
-        // mirrors the existing in-memory behaviour for collection refs and is what consumers see
-        // when they add @Aggregate without arguments.
-        val itemOnDelete = agg.onDeleteName
-
-        file.write(
-            buildString {
-                if (packageName.isNotEmpty()) {
-                    appendLine("package $packageName")
-                    appendLine()
-                }
-                appendLine("import net.transgressoft.lirp.entity.CascadeAction")
-                appendLine("import net.transgressoft.lirp.persistence.ColumnType")
-                appendLine("import net.transgressoft.lirp.persistence.sql.JunctionColumnDef")
-                appendLine("import net.transgressoft.lirp.persistence.sql.JunctionTableDef")
-                appendLine()
-                appendLine("/** KSP-generated junction table descriptor for $parentSimpleName.${agg.propertyName} → $itemSimpleName. */")
-                appendLine("public object $descriptorName : JunctionTableDef {")
-                appendLine("    override val tableName: String = \"$junctionTableName\"")
-                appendLine("    override val parentTableName: String = \"$parentTableName\"")
-                appendLine("    override val itemTableName: String = \"$itemTableName\"")
-                appendLine("    override val isOrdered: Boolean = ${agg.isOrdered}")
-                appendLine("    override val parentFkOnDelete: CascadeAction = CascadeAction.CASCADE")
-                appendLine("    override val itemFkOnDelete: CascadeAction = CascadeAction.$itemOnDelete")
-                appendLine("    override val columns: List<JunctionColumnDef> = listOf(")
-                appendLine("        JunctionColumnDef(name = \"parent_id\", type = $parentPkType, primaryKey = true),")
-                if (agg.isOrdered) {
-                    appendLine("        JunctionColumnDef(name = \"item_id\", type = $itemPkType, primaryKey = true),")
-                    appendLine("        JunctionColumnDef(name = \"position\", type = ColumnType.IntType)")
-                } else {
-                    appendLine("        JunctionColumnDef(name = \"item_id\", type = $itemPkType, primaryKey = true)")
-                }
-                appendLine("    )")
-                appendLine("}")
-            }.toByteArray()
-        )
-        file.close()
-
-        logger.info("Generated $packageName.$descriptorName for $parentSimpleName.${agg.propertyName}")
-    }
-
-    private fun pkColumnTypeExpression(classDecl: KSClassDeclaration): String? {
-        val idProp = classDecl.getAllProperties().firstOrNull { it.simpleName.asString() == "id" } ?: return null
-        return mapToColumnTypeExpression(idProp, persistenceAnnotation = null)
-    }
 }
-
-private fun String.toSnakeCase(): String =
-    replace(Regex("([a-z\\d])([A-Z])"), "$1_$2")
-        .replace(Regex("([A-Z]+)([A-Z][a-z])"), "$1_$2")
-        .lowercase()
-
-internal data class ColumnMeta(
-    val name: String,
-    val propertyName: String,
-    val typeExpression: String,
-    val typeFqn: String,
-    val nullable: Boolean,
-    val isPrimaryKey: Boolean,
-    val isEnum: Boolean = false,
-    val isMutable: Boolean = false,
-    val isCtorParam: Boolean = false,
-    val isVersion: Boolean = false,
-    val converterFqn: String? = null,
-    val converterSqlFqn: String? = null,
-    /**
-     * Dot-separated entity-access path from the entity root down to the underlying scalar.
-     * For top-level scalars this equals [propertyName]; for fields produced by flattening an
-     * `@Embedded` value object this is the nested path (e.g. `album.performer.name`). Used by
-     * `toParams` to dereference the value when binding the column.
-     */
-    val embeddedPath: String = propertyName,
-    /**
-     * `true` when this column was produced by flattening an `@Embedded` value object. Embedded-
-     *  derived columns are excluded from the entity-level mutability gate and from `applyRow`
-     *  because embeddables are reconstructed wholesale via the primary constructor on each
-     *  `fromRow` call.
-     */
-    val isInsideEmbedded: Boolean = false
-)
-
-private data class AggregatePropertyMeta(
-    val property: KSPropertyDeclaration,
-    val propertyName: String,
-    val isCollection: Boolean,
-    val isOrdered: Boolean,
-    val onDeleteName: String,
-    val referencedClass: KSClassDeclaration,
-    val backingScalarName: String?,
-    val backingCollectionName: String? = null
-)
-
-private data class JunctionRefInfo(
-    val propertyName: String,
-    val backingFieldName: String,
-    val junctionObjectName: String,
-    val isOrdered: Boolean,
-    val itemKeyTypeSimpleName: String,
-    val isMutableList: Boolean
-)
-
-private data class ForeignKeyMeta(
-    val columnName: String,
-    val referencedTable: String,
-    val referencedColumn: String,
-    val onDelete: String
-)
